Add search component in TakeOrder view
